refactor(profile): 프로필 도메인 파일 구조 정리
profile.* 네이밍 통일, 단일 사용 상수 colocate, ProfileSectionStream 추가 및 types 폴더 제거.
