+++ features/profile/api/fetch-my-profile.ts
@@ -1,5 +1,5 @@
 import { apiRequest } from '@/shared/lib/api';
 
-import type { Profile } from '../types';
+import type { Profile } from '../profile.types';
 
 export const fetchMyProfile = (): Promise<Profile> => apiRequest('/users/profile');
