+++ features/profile/api/fetch-profile-by-id.ts
@@ -1,5 +1,5 @@
 import { apiRequest } from '@/shared/lib/api';
 
-import type { Profile } from '../types';
+import type { Profile } from '../profile.types';
 
 export const fetchProfileById = (userId: number): Promise<Profile> => apiRequest<Profile>(`/users/profile/${userId}`);
