+++ features/profile/components/profile-edit-form.tsx
@@ -10,11 +10,14 @@ import { Button } from '@/shared/components/button';
 import { TextField } from '@/shared/components/text-field';
 import { TextareaField } from '@/shared/components/textarea-field';
 
-import type { Profile, UpdateProfilePayload } from '../types';
+import type { Profile, UpdateProfilePayload } from '../profile.types';
 
-import { PROFILE_BIO_MAX_LENGTH, PROFILE_IMAGE_ACCEPT, PROFILE_NICKNAME_MAX_LENGTH } from '../constants';
+import { PROFILE_BIO_MAX_LENGTH, PROFILE_NICKNAME_MAX_LENGTH } from '../profile.constants';
 import { ProfileAvatar } from './profile-avatar';
 import { useProfileEditForm } from '../hooks/use-profile-edit-form';
+
+/** 프로필 이미지 file input accept 속성 */
+const PROFILE_IMAGE_ACCEPT = 'image/png,image/jpeg,image/webp';
 
 type ProfileEditFormProps = {
   profile: Profile;
