--- features/profile/constants.ts
@@ -1,5 +1,0 @@
-export const PROFILE_NICKNAME_MAX_LENGTH = 20;
-export const PROFILE_BIO_MAX_LENGTH = 50;
-export const PROFILE_IMAGE_MAX_SIZE = 5 * 1024 * 1024;
-export const PROFILE_IMAGE_ACCEPT = 'image/png,image/jpeg,image/webp';
-export const DEFAULT_PROFILE_IMAGE_SRC = '/default-profile.png';
