+++ features/profile/hooks/use-profile-by-id.query.ts
@@ -1,7 +1,7 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 import { fetchProfileById } from '../api/fetch-profile-by-id';
-import { profileKeys } from '../query-keys';
+import { profileKeys } from '../profile.query-keys';
 
 export const useProfileById = (userId: number) =>
   useSuspenseQuery({
