+++ features/profile/hooks/use-update-profile.mutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { updateProfile } from '../api/update-profile';
-import { profileKeys } from '../query-keys';
+import { profileKeys } from '../profile.query-keys';
 
 export const useUpdateProfile = () => {
   const queryClient = useQueryClient();
