+++ features/profile/profile.constants.ts
@@ -1,0 +1,5 @@
+/** 닉네임 최대 글자 수 — 스키마 검증과 입력 필드 maxLength에서 공유 */
+export const PROFILE_NICKNAME_MAX_LENGTH = 20;
+
+/** 한 줄 소개 최대 글자 수 — 스키마 검증과 입력 필드 maxLength에서 공유 */
+export const PROFILE_BIO_MAX_LENGTH = 50;
